Fix ArgumentOutOfRangeException in WebSearch Plugin (#4041)

+++ Flow.Launcher/SettingPages/ViewModels/SettingsPaneHotkeyViewModel.cs
@@ -79,9 +79,12 @@ public partial class SettingsPaneHotkeyViewModel : BaseModel
         if (window.ShowDialog() is not true) return;
 
         var index = Settings.CustomPluginHotkeys.IndexOf(settingItem);
-        Settings.CustomPluginHotkeys[index] = new CustomPluginHotkey(window.Hotkey, window.ActionKeyword);
-        HotKeyMapper.RemoveHotkey(settingItem.Hotkey); // remove origin hotkey
-        HotKeyMapper.SetCustomQueryHotkey(Settings.CustomPluginHotkeys[index]); // set new hotkey
+        if (index >= 0 && index < Settings.CustomPluginHotkeys.Count)
+        {
+            Settings.CustomPluginHotkeys[index] = new CustomPluginHotkey(window.Hotkey, window.ActionKeyword);
+            HotKeyMapper.RemoveHotkey(settingItem.Hotkey); // remove origin hotkey
+            HotKeyMapper.SetCustomQueryHotkey(Settings.CustomPluginHotkeys[index]); // set new hotkey
+        }
     }
 
     [RelayCommand]
@@ -140,7 +143,10 @@ public partial class SettingsPaneHotkeyViewModel : BaseModel
         if (window.ShowDialog() is not true) return;
 
         var index = Settings.CustomShortcuts.IndexOf(settingItem);
-        Settings.CustomShortcuts[index] = new CustomShortcutModel(window.Key, window.Value);
+        if (index >= 0 && index < Settings.CustomShortcuts.Count)
+        {
+            Settings.CustomShortcuts[index] = new CustomShortcutModel(window.Key, window.Value);
+        }
     }
 
     [RelayCommand]
